+++ packages/test-helpers/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    preset: 'ts-jest',
+    transform: {
+      '^.+\\.(ts|js)?$': 'ts-jest',
+    },
+    coverageReporters: ['json', 'lcov', 'text-summary'],
+    collectCoverageFrom: ['src/**/*.js', 'src/**/*.ts'],
+    testEnvironment : 'jsdom'
+  };
+
